Deleted @NoArgsConstructor by mistake in User.java before uploading
Before uploading the files i did some last minute removal of extra comment lines in which  i deleted @NoArgsConstructor line by mistake.... in my User Entity class....... i noticed it now while making some changes in the program to test and make it run for h2 database. Really Sorry about that and the inconvenience caused. Kindly consider the request of considering the  newly added User.java for for evaluation of the project.